added MapNodes for temporal filtering and percent signal change; placed percent signal change functionality in new directory called conversion
