+++ spynoza/utils.py
@@ -1,3 +1,4 @@
+import nipype.pipeline as pe
 from nipype.interfaces.utility import Function
 import numpy as np
 
@@ -224,58 +225,6 @@ def pickfirst(files):
             return files
     else:
         return files
-
-
-def percent_signal_change(in_file, func='mean'):
-    """Converts data in a nifti-file to percent signal change.
-
-    Takes a 4D fMRI nifti-file and subtracts the
-    mean data from the original data, after which division
-    by the mean or median and multiplication with 100.
-
-    Parameters
-    ----------
-    in_file : str
-        Absolute path to nifti-file.
-    func : string ['mean', 'median'] (default: 'mean')
-        the function used to calculate the first moment
-
-    Returns
-    -------
-    out_file : str
-        Absolute path to converted nifti-file.
-    """
-
-    import nibabel as nib
-    import numpy as np
-    import os
-    import bottleneck as bn
-
-    data = nib.load(in_file)
-    dims = data.shape
-    affine = data.affine
-    header = data.header
-
-    if func == 'mean':
-        data_m = bn.nanmean(data.get_data(), axis=-1)
-    elif func == 'median':
-        data_m = bn.nanmedian(data.get_data(), axis=-1)
-
-    data_psc = (100.0 * (np.nan_to_num(data.get_data()).transpose(
-        (3, 0, 1, 2)) - data_m) / data_m).transpose((1, 2, 3, 0))
-    img = nib.Nifti1Image(np.nan_to_num(data_psc), affine=affine, header=header)
-
-    new_name = os.path.basename(in_file).split('.')[:-2][0] + '_psc.nii.gz'
-    out_file = os.path.abspath(new_name)
-    nib.save(img, out_file)
-
-    return out_file
-
-
-Percent_signal_change = Function(function=percent_signal_change,
-                                 input_names=['in_file', 'func'],
-                                 output_names=['out_file'])
-
 
 def average_over_runs(in_files, func='mean', output_filename=None):
     """Converts data in a nifti-file to percent signal change.
